+++ actix-web/CHANGES.md
@@ -6,7 +6,8 @@
 
 - Add `HttpServer::{bind, listen}_auto_h2c()` method behind new `http2` crate feature.
 - Add `Resource::{get, post, etc...}` methods for more concisely adding routes that don't need additional guards.
-- Add several missing convenience methods on `HttpResponse` for their respective status codes.
+- Add `web::Payload::to_bytes[_limited]()` helper methods.
+- Add missing constructors on `HttpResponse` for several status codes.
 
 ### Changed